Add more redirects for HTML-spec URLs
This change adds a large number of redirects for HTML-spec URLs. They’re
mostly for handling pathnames for files that were part of W3C HTML forks
but for which there aren’t (now) any files with the same names.

Otherwise, without adding these redirects, we are getting a large number
of 404s logged for broken links coming in from those W3C redirects.

The set of redirects added here was put together based on
analyzing/grepping the nginx logs for 404s.